refactor(sessions): move sessions view inside its own folder
