+++ src/components/SessionsView/SessionsViewRow.jsx
